Fixed newline characters at the end of raw files. Added a new dataclass Adjustments to group all adjustments together. (#336)
