+++ reconstruction/ecoli/flat/adjustments/protein_deg_rates_adjustments.tsv
