+++ reconstruction/ecoli/flat/adjustments/rna_deg_rates_adjustments.tsv
