+++ reconstruction/ecoli/flat/base_codes/amino_acids.tsv
